Fix casting issue

+++ Fastor/simd_vector/extintrin.h
@@ -847,7 +847,7 @@ FASTOR_INLINE double sqrts(double a) {return _mm_cvtsd_f64(_mm_sqrt_pd(_mm_set1_
 
 // helper functions for going from array to mask and vice-versa
 // used when AVX512 masking is available
-template <size_t N, enable_if_t_<N==2 || N==4 || N==8,bool> = false>
+template <int N, enable_if_t_<N==2 || N==4 || N==8,bool> = false>
 inline uint8_t array_to_mask(const int (&b)[N])
 {
     uint8_t c = 0;
@@ -858,7 +858,7 @@ inline uint8_t array_to_mask(const int (&b)[N])
     }
     return c;
 }
-template <size_t N, enable_if_t_<N==16,bool> = false>
+template <int N, enable_if_t_<N==16,bool> = false>
 inline uint16_t array_to_mask(const int (&b)[N])
 {
     uint16_t c = 0;
@@ -869,7 +869,7 @@ inline uint16_t array_to_mask(const int (&b)[N])
     }
     return c;
 }
-template <size_t N, enable_if_t_<N==32,bool> = false>
+template <int N, enable_if_t_<N==32,bool> = false>
 inline uint32_t array_to_mask(const int (&b)[N])
 {
     uint32_t c = 0;
@@ -880,7 +880,7 @@ inline uint32_t array_to_mask(const int (&b)[N])
     }
     return c;
 }
-template <size_t N, enable_if_t_<N==64,bool> = false>
+template <int N, enable_if_t_<N==64,bool> = false>
 inline uint64_t array_to_mask(const int (&b)[N])
 {
     uint64_t c = 0;
@@ -893,7 +893,7 @@ inline uint64_t array_to_mask(const int (&b)[N])
 }
 
 
-template <size_t N, enable_if_t_<N==2 || N==4 || N==8,bool> = false>
+template <int N, enable_if_t_<N==2 || N==4 || N==8,bool> = false>
 inline void mask_to_array(uint8_t c, int (&b)[N])
 {
     for (int i=0; i < N; ++i)
@@ -903,7 +903,7 @@ inline void mask_to_array(uint8_t c, int (&b)[N])
     for (int i=0; i < N; ++i)
         b[i] *= -1;
 }
-template <size_t N, enable_if_t_<N==16,bool> = false>
+template <int N, enable_if_t_<N==16,bool> = false>
 inline void mask_to_array(uint16_t c, int (&b)[N])
 {
     for (int i=0; i < N; ++i)
@@ -913,7 +913,7 @@ inline void mask_to_array(uint16_t c, int (&b)[N])
     for (int i=0; i < N; ++i)
         b[i] *= -1;
 }
-template <size_t N, enable_if_t_<N==32,bool> = false>
+template <int N, enable_if_t_<N==32,bool> = false>
 inline void mask_to_array(uint32_t c, int (&b)[N])
 {
     for (int i=0; i < N; ++i)
@@ -923,7 +923,7 @@ inline void mask_to_array(uint32_t c, int (&b)[N])
     for (int i=0; i < N; ++i)
         b[i] *= -1;
 }
-template <size_t N, enable_if_t_<N==64,bool> = false>
+template <int N, enable_if_t_<N==64,bool> = false>
 inline void mask_to_array(uint64_t c, int (&b)[N])
 {
     for (int i=0; i < N; ++i)
